Move `AroundLock` class to separate header file
